Move cpp files to crypto/starkware package
