+++ core/crypto/starkware/crypto.go
@@ -1,4 +1,4 @@
-package crypto
+package starkware
 
 import (
 	"encoding/hex"
