+++ core/crypto/starkware/crypto_test.go
@@ -1,4 +1,4 @@
-package crypto
+package starkware
 
 import "testing"
